Rename apps/src/templates/census2017 directory to not have a year
