+++ apps/src/sites/code.org/pages/public/yourschool/thankyou.js
@@ -6,7 +6,7 @@ import isRtl from '@cdo/apps/code-studio/isRtlRedux';
 import initResponsive from '@cdo/apps/code-studio/responsive';
 import responsive from '@cdo/apps/code-studio/responsiveRedux';
 import {getStore, registerReducers} from '@cdo/apps/redux';
-import YourSchoolResources from '@cdo/apps/templates/census2017/YourSchoolResources';
+import YourSchoolResources from '@cdo/apps/templates/census/YourSchoolResources';
 
 registerReducers({isRtl, responsive});
