+++ apps/src/templates/census/CensusTeacherBanner.jsx
