+++ apps/src/templates/census/SchoolAutocompleteDropdownWithLabel.jsx
